requirements.txt
Pillow
pylibdmtx
pyzxing
opencv-python==4.5.1.48
